requirements.txt
datasets
transformers
scikit-learn
pandas
matplotlib
seaborn
torch
groq
python-dotenv